requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
groq>=0.12.0
tavily-python>=0.5.0
python-dotenv>=1.0.1
pydantic>=2.10.3
websockets>=13.1
httpx>=0.28.0
psutil>=6.1.0
aiofiles>=23.2.1